Move errors and validate index pattern ⇒ NP (#51805)
* Move errors and validate index pattern to NP

* removed unused mock

* remvoed irrelevant mock

* Removed unneeded mocks

* fix test
