+++ src/legacy/core_plugins/data/public/index_patterns/index_patterns/index_patterns.test.ts
@@ -25,10 +25,6 @@ import {
   HttpServiceBase,
 } from 'kibana/public';
 
-jest.mock('../errors', () => ({
-  IndexPatternMissingIndices: jest.fn(),
-}));
-
 jest.mock('./index_pattern', () => {
   class IndexPattern {
     init = async () => {
