+++ src/legacy/ui/public/indices/validate/validate_index.test.js
@@ -18,7 +18,6 @@
  */
 
 jest.mock('ui/new_platform');
-jest.mock('ui/index_patterns');
 
 import { INDEX_ILLEGAL_CHARACTERS_VISIBLE } from '../constants';
