+++ src/plugins/data/public/index.ts
@@ -30,6 +30,7 @@ export * from '../common';
 
 export * from './autocomplete_provider';
 export * from './field_formats_provider';
+export * from './index_patterns';
 
 export * from './types';
