+++ src/plugins/data/public/index_patterns/lib/validate_index_pattern.test.ts
@@ -1,0 +1,42 @@
+/*
+ * Licensed to Elasticsearch B.V. under one or more contributor
+ * license agreements. See the NOTICE file distributed with
+ * this work for additional information regarding copyright
+ * ownership. Elasticsearch B.V. licenses this file to you under
+ * the Apache License, Version 2.0 (the "License"); you may
+ * not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *    http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing,
+ * software distributed under the License is distributed on an
+ * "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
+ * KIND, either express or implied.  See the License for the
+ * specific language governing permissions and limitations
+ * under the License.
+ */
+
+import { CONTAINS_SPACES_KEY, ILLEGAL_CHARACTERS_KEY, ILLEGAL_CHARACTERS_VISIBLE } from './types';
+
+import { validateIndexPattern } from './validate_index_pattern';
+
+describe('Index Pattern Utils', () => {
+  describe('Validation', () => {
+    it('should not allow space in the pattern', () => {
+      const errors = validateIndexPattern('my pattern');
+      expect(errors[CONTAINS_SPACES_KEY]).toBe(true);
+    });
+
+    it('should not allow illegal characters', () => {
+      ILLEGAL_CHARACTERS_VISIBLE.forEach(char => {
+        const errors = validateIndexPattern(`pattern${char}`);
+        expect(errors[ILLEGAL_CHARACTERS_KEY]).toEqual([char]);
+      });
+    });
+
+    it('should return empty object when there are no errors', () => {
+      expect(validateIndexPattern('my-pattern-*')).toEqual({});
+    });
+  });
+});
