+++ src/plugins/data/public/index_patterns/lib/validate_index_pattern.ts
@@ -1,0 +1,51 @@
+/*
+ * Licensed to Elasticsearch B.V. under one or more contributor
+ * license agreements. See the NOTICE file distributed with
+ * this work for additional information regarding copyright
+ * ownership. Elasticsearch B.V. licenses this file to you under
+ * the Apache License, Version 2.0 (the "License"); you may
+ * not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *    http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing,
+ * software distributed under the License is distributed on an
+ * "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
+ * KIND, either express or implied.  See the License for the
+ * specific language governing permissions and limitations
+ * under the License.
+ */
+
+import { ILLEGAL_CHARACTERS_VISIBLE, CONTAINS_SPACES_KEY, ILLEGAL_CHARACTERS_KEY } from './types';
+
+function indexPatternContainsSpaces(indexPattern: string): boolean {
+  return indexPattern.includes(' ');
+}
+
+function findIllegalCharacters(indexPattern: string): string[] {
+  const illegalCharacters = ILLEGAL_CHARACTERS_VISIBLE.reduce((chars: string[], char: string) => {
+    if (indexPattern.includes(char)) {
+      chars.push(char);
+    }
+    return chars;
+  }, []);
+
+  return illegalCharacters;
+}
+
+export function validateIndexPattern(indexPattern: string) {
+  const errors: Record<string, any> = {};
+
+  const illegalCharacters = findIllegalCharacters(indexPattern);
+
+  if (illegalCharacters.length) {
+    errors[ILLEGAL_CHARACTERS_KEY] = illegalCharacters;
+  }
+
+  if (indexPatternContainsSpaces(indexPattern)) {
+    errors[CONTAINS_SPACES_KEY] = true;
+  }
+
+  return errors;
+}
