+++ x-pack/legacy/plugins/cross_cluster_replication/__jest__/client_integration/follower_indices_list.test.js
@@ -9,7 +9,6 @@ import { setupEnvironment, pageHelpers, nextTick, getRandomString } from './help
 import { getFollowerIndexMock } from '../../fixtures/follower_index';
 
 jest.mock('ui/new_platform');
-jest.mock('ui/index_patterns');
 
 jest.mock('ui/chrome', () => ({
   addBasePath: () => 'api/cross_cluster_replication',
