+++ x-pack/legacy/plugins/cross_cluster_replication/__jest__/client_integration/home.test.js
@@ -8,7 +8,6 @@
 import { setupEnvironment, pageHelpers, nextTick } from './helpers';
 
 jest.mock('ui/new_platform');
-jest.mock('ui/index_patterns');
 
 const { setup } = pageHelpers.home;
