+++ x-pack/legacy/plugins/cross_cluster_replication/public/app/services/auto_follow_pattern_validators.test.js
@@ -8,7 +8,6 @@
 import { validateAutoFollowPattern } from './auto_follow_pattern_validators';
 
 jest.mock('ui/new_platform');
-jest.mock('ui/index_patterns');
 
 describe('Auto-follow pattern validators', () => {
   describe('validateAutoFollowPattern()', () => {
