+++ x-pack/legacy/plugins/rollup/__jest__/client_integration/job_create_date_histogram.test.js
@@ -9,7 +9,6 @@ import moment from 'moment-timezone';
 import { setupEnvironment, pageHelpers } from './helpers';
 
 jest.mock('ui/new_platform');
-jest.mock('ui/index_patterns');
 
 jest.mock('lodash/function/debounce', () => fn => fn);
