+++ x-pack/legacy/plugins/rollup/__jest__/client_integration/job_create_terms.test.js
@@ -7,7 +7,6 @@
 import { setupEnvironment, pageHelpers } from './helpers';
 
 jest.mock('ui/new_platform');
-jest.mock('ui/index_patterns');
 
 jest.mock('lodash/function/debounce', () => fn => fn);
